Refactor: move socket creation and configuration to separate method
This commit refactors the initialization of the WebSocket interface in the `src/index.ts` file. It moves the creation and configuration of the socket to a separate method, which is then called twice: once when the SIP socket is created, and once when the UA is initialized.

Additionally, a new method `getSipStatus()` is added to the `src/sip.ts` file to retrieve the current status of the SIP socket. This method is used in the `index.html` file to display the status of the SIP connection.

In the `package.json` file, the `prettier` package is added to the dependencies. The `prettierrc` file is added to the repository to configure Prettier.

Finally, the `pnpm-lock.yaml` file is updated to include the new dependencies.